+++ package.json
@@ -13,6 +13,7 @@
     "test": "echo \"Write tests! -> https://gatsby.dev/unit-testing\" && exit 1"
   },
   "dependencies": {
+    "@geedmo/yamm": "^4.0.0",
     "apollo-server-express": "2.9.12",
     "aws-sdk": "^2.577.0",
     "babel-plugin-styled-components": "^1.10.6",
@@ -29,11 +30,12 @@
     "gatsby": "^2.17.6",
     "gatsby-plugin-google-fonts": "^1.0.1",
     "gatsby-plugin-react-helmet": "^3.1.13",
-    "gatsby-plugin-sass": "^2.1.23",
+    "gatsby-plugin-sass": "^2.1.24",
     "gatsby-plugin-styled-components": "^3.1.11",
     "gatsby-source-graphql": "^2.1.24",
     "graphql": "^14.5.8",
     "graphql-tools": "^4.0.6",
+    "node-sass": "^4.13.0",
     "prop-types": "^15.7.2",
     "react": "^16.11.0",
     "react-bootstrap": "^1.0.0-beta.14",
